refactor: rename push/ → web-push/, move client.ts to provider root
Provider dir named for the external system (web-push protocol).
client.ts at provider root per convention, not buried in lib/.
